pyproject.toml
[tool.poetry]
name = "miniproject1"
version = "0.1.0"
description = ""
authors = ["desylosanticho <desylo.santicho@prosa.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.4.1"
cffi = "1.15.0"
cryptography = "2.8"
Django = "3.2.8"
django-rest-knox = "4.1.0"
djangorestframework = "3.12.4"
pycparser = "2.20"
pytz = "2021.3"
six = "1.16.0"
sqlparse = "0.4.2"
typing-extensions = "3.10.0.2"
opencv-python = "^4.5.5"
Pillow = "^9.0.1"
django-extensions = "^3.1.5"
Markdown = "^3.3.6"
django-markdown-deux = "^1.0.5"
django-filter = "^21.1"
drf-proxy-pagination = "^0.2.0"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
